changed sizes to be doubles

+++ src/graphConverter.cc
@@ -169,14 +169,14 @@ void print_analysis(vector< pair<node_t, node_t> > edges, node_t N, edge_t M){
   printf("Tail verticies: %i\n", tail_verticies);
   printf("Head verticies: %i\n", head_verticies);
   printf("Disconnected verticies: %i\n", disc_verticies);
-  float sparseSize = ((float)(N*max_degree*8))/((float)(1024*1024*1024));
+  double sparseSize = ((double)(N*max_degree*8))/((double)(1024*1024*1024));
   printf("Unragged sparse size: %fGB\n", sparseSize);
-  float denseSize = ((float)(N*N*4))/((float)(1024*1024*1024));
+  double denseSize = ((double)(N*N*4))/((double)(1024*1024*1024));
   printf("Dense size: %fGB\n", denseSize);
-  printf("Usable for GTX480 dense:  %s\n", (denseSize<1.53)?"yes":"no");
-  printf("Usable for GTX480 sparse: %s\n", (sparseSize<1.53)?"yes":"no");
-  printf("Usable for    K40 dense:  %s\n", (denseSize<11.5)?"yes":"no");
-  printf("Usable for    K40 sparse: %s\n", (sparseSize<11.5)?"yes":"no");
+  printf("Usable for GTX480 dense:  %s\n", (denseSize<(double)1.53)?"yes":"no");
+  printf("Usable for GTX480 sparse: %s\n", (sparseSize<(double)1.53)?"yes":"no");
+  printf("Usable for    K40 dense:  %s\n", (denseSize<(double)11.5)?"yes":"no");
+  printf("Usable for    K40 sparse: %s\n", (sparseSize<(double)11.5)?"yes":"no");
   printf("\n\n\n");
 
   delete[] deg;
